# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Asrar Everyday API",
    "version": "1.0.0",
    "description": "RESTful API for Islamic spiritual calculations based on ʿIlm al-Ḥurūf (Science of Letters). Provides endpoints for name destiny analysis, spiritual guidance (Istikhara), and relationship compatibility assessment.",
    "contact": {
      "name": "Asrar Everyday Support",
      "email": "support@asrar-everyday.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://asrar-everyday.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.asrar-everyday.com/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/compatibility": {
      "post": {
        "tags": ["Compatibility"],
        "summary": "Calculate Relationship Compatibility",
        "description": "Analyzes compatibility between two individuals using three methods (Spiritual-Destiny/Mod-9, Elemental-Temperament/Mod-4, Planetary-Cosmic/Mod-7). When mother names are provided, performs advanced four-layer analysis examining different relationship dimensions.",
        "operationId": "calculateCompatibility",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompatibilityRequest"
              },
              "examples": {
                "basic": {
                  "summary": "Basic 3-method compatibility",
                  "value": {
                    "person1Name": "Ahmad",
                    "person2Name": "Khadija",
                    "language": "en",
                    "abjadSystem": "maghribi"
                  }
                },
                "fourLayer": {
                  "summary": "Four-layer deep analysis",
                  "value": {
                    "person1Name": "Hassan",
                    "person1MotherName": "Zainab",
                    "person2Name": "Fatima",
                    "person2MotherName": "Maryam",
                    "language": "en",
                    "abjadSystem": "maghribi"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful compatibility analysis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompatibilityResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "person1": {
                      "name": "Hassan",
                      "arabicName": "حسن"
                    },
                    "person2": {
                      "name": "Fatima",
                      "arabicName": "فاطمة"
                    },
                    "basicMethods": {
                      "spiritualDestiny": {
                        "score": 78,
                        "quality": "excellent"
                      },
                      "elementalTemperament": {
                        "score": 65,
                        "quality": "good"
                      },
                      "planetaryCosmic": {
                        "score": 54,
                        "quality": "moderate"
                      }
                    },
                    "overallScore": 66,
                    "compatibilityQuality": "good",
                    "recommendation": "Strong foundation with good prospects. Work on communication."
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CompatibilityRequest": {
        "type": "object",
        "required": ["person1Name", "person2Name", "language"],
        "properties": {
          "person1Name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Ahmad"
          },
          "person1MotherName": {
            "type": "string",
            "maxLength": 100,
            "description": "For four-layer analysis",
            "example": "Zainab"
          },
          "person2Name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Khadija"
          },
          "person2MotherName": {
            "type": "string",
            "maxLength": 100,
            "description": "For four-layer analysis",
            "example": "Maryam"
          },
          "language": {
            "type": "string",
            "enum": ["en", "fr", "ar"],
            "example": "en"
          },
          "abjadSystem": {
            "type": "string",
            "enum": ["maghribi", "mashriqi"],
            "default": "maghribi",
            "example": "maghribi"
          }
        }
      },
...
      "CompatibilityResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid input parameters"
              },
              "details": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error - invalid input parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Invalid input parameters",
                "details": {
                  "personName": "Name must be between 1 and 100 characters"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "INTERNAL_ERROR",
                "message": "An unexpected error occurred"
              }
            }
          }
        }
      }
    }
  }
}